training/requirements.txt
mlflow
pyyaml
pandas
datasets
transformers
evaluate
rouge_score
accelerate
torch
sentencepiece